requirements.txt
django >=1.11
django-mptt >= 0.9
pillow >= 2.1
